+++ src/components/Grid/index.ts
@@ -1,0 +1,2 @@
+export { default as OrigamGrid } from './OrigamGrid.vue'
+export { default as OrigamGridItem } from './OrigamGridItem.vue'